genai_kafka/ecomm/pyproject.toml
[tool.poetry]
name = "ecomm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "3.12.5"
requests = "^2.32.3"
confluent-kafka = "^2.6.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
kafka-python-ng = "^2.2.3"
protobuf = "3.20.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
